requirements.txt
opencv-python=='4.1.0'
Pillow=='6.1.0'